+++ ZendeskSell/LeadConversions/ILeadConversionActions.cs
@@ -6,7 +6,7 @@ using ZendeskSell.Models;
 
 namespace ZendeskSell.LeadConversions {
     public interface ILeadConversionActions {
-        Task<ZendeskSellCollectionResponse<LeadConversionResponse>> GetAsync(int pageNumber, int numPerPage, int? leadID = null, int? individualID = null, int? organizationID = null, int? dealID = null);
+        Task<ZendeskSellCollectionResponse<LeadConversionResponse>> GetAsync(int pageNumber, int numPerPage, long? leadID = null, long? individualID = null, long? organizationID = null, long? dealID = null);
         Task<ZendeskSellObjectResponse<LeadConversionResponse>> CreateAsync(LeadConversionRequest leadConversion);
     }
 }
