+++ ZendeskSell/LeadConversions/LeadConversion.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 namespace ZendeskSell.LeadConversions {
     public class LeadConversionRequest {
         [JsonProperty("lead_id")]
-        public int LeadID { get; set; }
+        public long LeadID { get; set; }
         [JsonProperty("owner_id")]
         public int? OwnerID { get; set; }
         [JsonProperty("create_deal")]
@@ -15,13 +15,13 @@ namespace ZendeskSell.LeadConversions {
         [JsonProperty("id")]
         public int ID { get; set; }
         [JsonProperty("lead_id")]
-        public int LeadID { get; set; }
+        public long LeadID { get; set; }
         [JsonProperty("individual_id")]
-        public int? IndividualID { get; set; }
+        public long? IndividualID { get; set; }
         [JsonProperty("organization_id")]
-        public int? OrganizationID { get; set; }
+        public long? OrganizationID { get; set; }
         [JsonProperty("deal_id")]
-        public int? DealID { get; set; }
+        public long? DealID { get; set; }
         [JsonProperty("created_at")]
         public DateTimeOffset CreatedAt { get; set; }
     }
