+++ ZendeskSell/LineItems/ILineItemActions.cs
@@ -4,9 +4,9 @@ using ZendeskSell.Models;
 namespace ZendeskSell.LineItems {
     public interface ILineItemActions {
         Task<ZendeskSellCollectionResponse<LineItemResponse>> GetAsync(int orderID, int pageNumber, int numPerPage);
-        Task<ZendeskSellObjectResponse<LineItemResponse>> GetOneAsync(int orderID, int lineItemID);
+        Task<ZendeskSellObjectResponse<LineItemResponse>> GetOneAsync(int orderID, long lineItemID);
         Task<ZendeskSellObjectResponse<LineItemResponse>> CreateAsync(int orderID, LineItemRequest lineItem);
         /// <returns><see langword="null"/> on success</returns>
-        Task<ZendeskSellDeleteResponse> DeleteAsync(int orderID, int lineItemID);
+        Task<ZendeskSellDeleteResponse> DeleteAsync(int orderID, long lineItemID);
     }
 }
